refactor: move order quote query function to its own file
