+++ tauri-app/src/lib/services/orderQuote.ts
@@ -1,0 +1,14 @@
+import { get } from 'svelte/store';
+import { invoke } from '@tauri-apps/api';
+import { rpcUrl, orderbookAddress, subgraphUrl } from '$lib/stores/settings';
+
+export async function batchOrderQuotes(
+  orderHashes: string[],
+): Promise<{ maxOutput: string; ratio: string }[]> {
+  return await invoke('batch_order_quotes', {
+    orderHashes,
+    subgraphUrl: get(subgraphUrl),
+    rpcUrl: get(rpcUrl),
+    orderbook: get(orderbookAddress),
+  });
+}
